Refactor gallery helpers and add unit tests
